Introduce a new CliCompiler
The only compiler available as far as now was the PhpServerCompiler. I'm
not sure why, but it was taking way too much time on our current
project (> 1h).

This commit adds a CliCompiler that just do the same as the former
compiler, but from the CLI. It takes less than 5 seconds to compile
~5000 files.

Unfortunately, as pointed out by jderusse in his review, not all files
can be compiled in the same process as composer because some of these
files are already loaded in memory, thus compiling them would cause:

>  "PHP Fatal error: Cannot declare class Foo, because the name is
>  already in use in /path/Bar.php on line 3"